+++ apx_sdk/src/fetch.rs
@@ -22,16 +22,12 @@ use thiserror::Error;
 use apx_core::{
     http_signatures::create::HttpSignatureError,
     media_type::sniff_media_type,
-    url::{
-        canonical::is_same_origin,
-        http_uri::is_same_http_origin,
-    },
+    url::canonical::is_same_origin,
 };
 
 use super::{
     agent::FederationAgent,
     authentication::{
-        verify_portable_object,
         AuthenticationError,
     },
     constants::{AP_MEDIA_TYPE, AS_MEDIA_TYPE},
@@ -175,20 +171,19 @@ fn extract_fragment(
 /// Options for `fetch_object`
 #[derive(Default)]
 pub struct FetchObjectOptions {
-    /// Skip origin and content type checks?
-    pub skip_verification: bool,
-    /// List of trusted origins for a FEP-ef61 collection
-    pub fep_ef61_trusted_origins: Vec<String>,
+    /// Skip content type check?
+    pub skip_content_type_verification: bool,
 }
 
-struct FetchedObject {
-    value: JsonValue,
-    location: Url,
+/// Return type of the `fetch_object` function
+pub struct FetchedObject {
+    pub value: JsonValue,
+    pub location: Url,
 }
 
 impl FetchedObject {
     /// Verifies the origin of this object
-    fn verify_origin(&self) -> Result<(), FetchError> {
+    pub fn verify_origin(&self) -> Result<(), FetchError> {
         let object_id = self.value["id"].as_str()
             .ok_or(FetchError::NoObjectId(self.location.to_string()))?;
         let is_trusted = is_same_origin(self.location.as_str(), object_id)
@@ -199,7 +194,8 @@ impl FetchedObject {
         Ok(())
     }
 
-    fn extract_fragment(&self) -> Result<JsonValue, FetchError> {
+    /// Extracts the fragment if location URL contains a fragment ID
+    pub fn extract_fragment(&self) -> Result<JsonValue, FetchError> {
         let object_id = self.value["id"].as_str()
             .ok_or(FetchError::NoObjectId(self.location.to_string()))?;
         if let Some(fragment_id) = self.location.fragment() {
@@ -215,12 +211,13 @@ impl FetchedObject {
     }
 }
 
-/// Sends GET request to fetch ActivityPub object. Supports fragment resolution.
+/// Sends a GET request to fetch an ActivityPub object.  
+/// Does not perform authentication.
 pub async fn fetch_object(
     agent: &FederationAgent,
     object_url: &str,
     options: FetchObjectOptions,
-) -> Result<JsonValue, FetchError> {
+) -> Result<FetchedObject, FetchError> {
     // Don't follow redirects automatically,
     // because request needs to be signed again after every redirect
     let client = create_fetcher_client(
@@ -279,7 +276,7 @@ pub async fn fetch_object(
         AS_MEDIA_TYPE,
         "application/ld+json",
     ];
-    if !options.skip_verification && !ALLOWED_TYPES.contains(&content_type.as_str()) {
+    if !options.skip_content_type_verification && !ALLOWED_TYPES.contains(&content_type.as_str()) {
         return Err(FetchError::UnexpectedContentType(content_type));
     };
 
@@ -289,38 +286,7 @@ pub async fn fetch_object(
     let object_json: JsonValue = serde_json::from_slice(&object_bytes)
         .map_err(|_| FetchError::JsonParseError(object_location.to_string()))?;
     let object = FetchedObject { value: object_json, location: object_location };
-
-    if options.skip_verification {
-        let object_json = object.extract_fragment()?;
-        return Ok(object_json);
-    };
-
-    // Perform authentication
-    match verify_portable_object(&object.value) {
-        Ok(_) => (),
-        Err(AuthenticationError::InvalidObjectID(_)) => {
-            return Err(FetchError::UrlError);
-        },
-        Err(AuthenticationError::NotPortable) => {
-            // Verify origin if object is not portable
-            object.verify_origin()?;
-        },
-        Err(AuthenticationError::NoProof) => {
-            let is_trusted = options.fep_ef61_trusted_origins
-                .iter()
-                .any(|origin| {
-                    is_same_http_origin(object.location.as_str(), origin)
-                        .unwrap_or(false)
-                });
-            if !is_trusted {
-                return Err(FetchError::UnexpectedObjectId(object.location.to_string()));
-            };
-        },
-        Err(other_error) => return Err(FetchError::InvalidProof(other_error)),
-    };
-
-    let object_json = object.extract_fragment()?;
-    Ok(object_json)
+    Ok(object)
 }
 
 fn get_media_type(