+++ apx_sdk/CHANGELOG.md
@@ -9,10 +9,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Added `authority` field to `HttpSignatureData` struct.
+- Added `verify_fetched_object` function.
 
 ### Changed
 
 - Extract authority from request URI when processing `@authority` RFC-9421 component.
+- Do not perform authentication in `fetch_object` function.
+- Do not extract fragment in `fetch_object` function.
 
 ## [0.23.0] - 2026-05-05